+++ NEWS.md
@@ -1,4 +1,4 @@
-# codecheck 0.26.0.9000
+# codecheck 0.26.0.9001
 
 ## New Features